# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Live Tram API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/routes": {
      "get": {
        "tags": [
          "routes"
        ],
        "summary": "List all routes, optionally filtered by area or type",
        "operationId": "list_routes",
        "parameters": [
          {
            "name": "area_id",
            "in": "query",
            "description": "Filter by area ID",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "route_type",
            "in": "query",
            "description": "Filter by route type (e.g., \"tram\", \"bus\")",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of routes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/routes/{route_id}": {
      "get": {
        "tags": [
          "routes"
        ],
        "summary": "Get a single route with its stops",
        "operationId": "get_route",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "description": "Route OSM ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route details with stops",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDetail"
                }
              }
            }
          },
          "404": {
            "description": "Route not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/routes/{route_id}/geometry": {
      "get": {
        "tags": [
          "routes"
        ],
        "summary": "Get the geometry of a route as line segments",
        "operationId": "get_route_geometry",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "description": "Route OSM ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route geometry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteGeometry"
                }
              }
            }
          },
          "404": {
            "description": "Route not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stations": {
      "get": {
        "tags": [
          "stations"
        ],
        "summary": "List all stations that have platforms linked to them, optionally filtered by area",
        "operationId": "list_stations",
        "parameters": [
          {
            "name": "area_id",
            "in": "query",
            "description": "Filter by area ID",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of stations with their platforms and stop positions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Route": {
        "type": "object",
        "required": [
          "osm_id",
          "osm_type",
          "route_type"
        ],
        "properties": {
          "area_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "network": {
            "type": [
              "string",
              "null"
            ]
          },
          "operator": {
            "type": [
              "string",
              "null"
            ]
          },
          "osm_id": {
            "type": "integer",
            "format": "int64"
          },
          "osm_type": {
            "type": "string"
          },
          "ref": {
            "type": [
              "string",
              "null"
            ]
          },
          "route_type": {
            "type": "string"
          }
        }
      },
      "RouteDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Route"
          },
          {
            "type": "object",
            "required": [
              "stops"
            ],
            "properties": {
              "stops": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RouteStop"
                }
              }
            }
          }
        ]
      },
      "RouteGeometry": {
        "type": "object",
        "required": [
          "route_id",
          "segments"
        ],
        "properties": {
          "route_id": {
            "type": "integer",
            "format": "int64"
          },
          "segments": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "array",
                "items": {
                  "type": "number",
                  "format": "double"
                }
              }
            }
          }
        }
      },
      "RouteListResponse": {
        "type": "object",
        "required": [
          "routes"
        ],
        "properties": {
          "routes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Route"
            }
          }
        }
      },
      "RouteStop": {
        "type": "object",
        "required": [
          "sequence"
        ],
        "properties": {
          "platform_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "role": {
            "type": [
              "string",
              "null"
            ]
          },
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "station_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "station_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "stop_position_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "Station": {
        "type": "object",
        "required": [
          "osm_id",
          "osm_type",
          "lat",
          "lon",
          "platforms",
          "stop_positions"
        ],
        "properties": {
          "area_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "osm_id": {
            "type": "integer",
            "format": "int64"
          },
          "osm_type": {
            "type": "string"
          },
          "platforms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StationPlatform"
            }
          },
          "ref_ifopt": {
            "type": [
              "string",
              "null"
            ]
          },
          "stop_positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StationStopPosition"
            }
          }
        }
      },
      "StationListResponse": {
        "type": "object",
        "required": [
          "stations"
        ],
        "properties": {
          "stations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Station"
            }
          }
        }
      },
      "StationPlatform": {
        "type": "object",
        "description": "Platform info nested in station response",
        "required": [
          "osm_id",
          "lat",
          "lon"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "osm_id": {
            "type": "integer",
            "format": "int64"
          },
          "ref": {
            "type": [
              "string",
              "null"
            ]
          },
          "ref_ifopt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "StationStopPosition": {
        "type": "object",
        "description": "Stop position info nested in station response",
        "required": [
          "osm_id",
          "lat",
          "lon"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "osm_id": {
            "type": "integer",
            "format": "int64"
          },
          "platform_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "ref": {
            "type": [
              "string",
              "null"
            ]
          },
          "ref_ifopt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}