# server/src/api/stations/list.rs
use axum::{
    Json,
    extract::{Query, State},
    http::StatusCode,
};
use serde::{Deserialize, Serialize};
use sqlx::{FromRow, SqlitePool};
use std::collections::HashMap;
use utoipa::{IntoParams, ToSchema};

use crate::api::{ErrorResponse, internal_error};

/// Internal struct for database row
#[derive(Debug, FromRow)]
struct StationRow {
    pub osm_id: i64,
    pub osm_type: String,
    pub name: Option<String>,
    pub ref_ifopt: Option<String>,
    pub lat: f64,
    pub lon: f64,
    pub area_id: Option<i64>,
}

/// Platform info nested in station response
#[derive(Debug, Serialize, ToSchema)]
pub struct StationPlatform {
    pub osm_id: i64,
    pub name: Option<String>,
    #[serde(rename = "ref")]
    pub platform_ref: Option<String>,
    pub ref_ifopt: Option<String>,
    pub lat: f64,
    pub lon: f64,
}

/// Internal row struct for platform query
#[derive(Debug, FromRow)]
struct PlatformRow {
    station_id: i64,
    osm_id: i64,
    name: Option<String>,
    #[sqlx(rename = "ref")]
    platform_ref: Option<String>,
    ref_ifopt: Option<String>,
    lat: f64,
    lon: f64,
}

/// Stop position info nested in station response
#[derive(Debug, Serialize, ToSchema)]
pub struct StationStopPosition {
    pub osm_id: i64,
    pub name: Option<String>,
    #[serde(rename = "ref")]
    pub stop_ref: Option<String>,
    pub ref_ifopt: Option<String>,
    pub lat: f64,
    pub lon: f64,
    pub platform_id: Option<i64>,
}

/// Internal row struct for stop position query
#[derive(Debug, FromRow)]
struct StopPositionRow {
    station_id: i64,
    osm_id: i64,
    name: Option<String>,
    #[sqlx(rename = "ref")]
    stop_ref: Option<String>,
    ref_ifopt: Option<String>,
    lat: f64,
    lon: f64,
    platform_id: Option<i64>,
}

#[derive(Debug, Serialize, ToSchema)]
pub struct Station {
    pub osm_id: i64,
    pub osm_type: String,
    pub name: Option<String>,
    pub ref_ifopt: Option<String>,
    pub lat: f64,
    pub lon: f64,
    pub area_id: Option<i64>,
    pub platforms: Vec<StationPlatform>,
    pub stop_positions: Vec<StationStopPosition>,
}

#[derive(Debug, Serialize, ToSchema)]
pub struct StationListResponse {
    pub stations: Vec<Station>,
}

#[derive(Debug, Deserialize, IntoParams)]
pub struct StationQuery {
    /// Filter by area ID
    pub area_id: Option<i64>,
}

/// List all stations that have platforms linked to them, optionally filtered by area
#[utoipa::path(
    get,
    path = "/api/stations",
    params(StationQuery),
    responses(
        (status = 200, description = "List of stations with their platforms and stop positions", body = StationListResponse),
        (status = 500, description = "Internal server error", body = ErrorResponse)
    ),
    tag = "stations"
)]
pub async fn list_stations(
    State(pool): State<SqlitePool>,
    Query(query): Query<StationQuery>,
) -> Result<Json<StationListResponse>, (StatusCode, Json<ErrorResponse>)> {
    // Only return stations that have at least one platform linked to them
    // This filters out bus-only stop_areas when we only have tram platforms
    let station_rows: Vec<StationRow> = if let Some(area_id) = query.area_id {
        sqlx::query_as(
            r#"
            SELECT DISTINCT s.osm_id, s.osm_type, s.name, s.ref_ifopt, s.lat, s.lon, s.area_id
            FROM stations s
            INNER JOIN platforms p ON p.station_id = s.osm_id
            WHERE s.area_id = ?
            ORDER BY s.name
            "#,
        )
        .bind(area_id)
        .fetch_all(&pool)
        .await
    } else {
        sqlx::query_as(
            r#"
            SELECT DISTINCT s.osm_id, s.osm_type, s.name, s.ref_ifopt, s.lat, s.lon, s.area_id
            FROM stations s
            INNER JOIN platforms p ON p.station_id = s.osm_id
            ORDER BY s.name
            "#,
        )
        .fetch_all(&pool)
        .await
    }
    .map_err(internal_error)?;

    if station_rows.is_empty() {
        return Ok(Json(StationListResponse { stations: vec![] }));
    }

    // Collect station IDs for batch queries
    let station_ids: Vec<i64> = station_rows.iter().map(|s| s.osm_id).collect();

    // Fetch all platforms for these stations in one query
    let platform_rows: Vec<PlatformRow> = sqlx::query_as(
        r#"
        SELECT station_id, osm_id, name, ref, ref_ifopt, lat, lon
        FROM platforms
        WHERE station_id IN (SELECT value FROM json_each(?))
        ORDER BY ref, name
        "#,
    )
    .bind(serde_json::to_string(&station_ids).unwrap_or_default())
    .fetch_all(&pool)
    .await
    .map_err(internal_error)?;

    // Fetch all stop_positions for these stations in one query
    let stop_rows: Vec<StopPositionRow> = sqlx::query_as(
        r#"
        SELECT station_id, osm_id, name, ref, ref_ifopt, lat, lon, platform_id
        FROM stop_positions
        WHERE station_id IN (SELECT value FROM json_each(?))
        ORDER BY ref, name
        "#,
    )
    .bind(serde_json::to_string(&station_ids).unwrap_or_default())
    .fetch_all(&pool)
    .await
    .map_err(internal_error)?;

    // Group platforms and stop_positions by station_id
    let mut platforms_by_station: HashMap<i64, Vec<StationPlatform>> = HashMap::new();
    for row in platform_rows {
        platforms_by_station
            .entry(row.station_id)
            .or_default()
            .push(StationPlatform {
                osm_id: row.osm_id,
                name: row.name,
                platform_ref: row.platform_ref,
                ref_ifopt: row.ref_ifopt,
                lat: row.lat,
                lon: row.lon,
            });
    }

    let mut stops_by_station: HashMap<i64, Vec<StationStopPosition>> = HashMap::new();
    for row in stop_rows {
        stops_by_station
            .entry(row.station_id)
            .or_default()
            .push(StationStopPosition {
                osm_id: row.osm_id,
                name: row.name,
                stop_ref: row.stop_ref,
                ref_ifopt: row.ref_ifopt,
                lat: row.lat,
                lon: row.lon,
                platform_id: row.platform_id,
            });
    }

    // Build final response
    let stations = station_rows
        .into_iter()
        .map(|row| Station {
            osm_id: row.osm_id,
            osm_type: row.osm_type,
            name: row.name,
            ref_ifopt: row.ref_ifopt,
            lat: row.lat,
            lon: row.lon,
            area_id: row.area_id,
            platforms: platforms_by_station.remove(&row.osm_id).unwrap_or_default(),
            stop_positions: stops_by_station.remove(&row.osm_id).unwrap_or_default(),
        })
        .collect();

    Ok(Json(StationListResponse { stations }))
}